requirements.txt
#Make sure the following python dependencies are installed

#General workflow requirements
numpy
matplotlib
scipy
tensorflow-gpu==1.15

#toy dataset requirements
mnist
sk-video
ffmpeg

#model requirement
opencv-python
scikit-image
h5py
